+++ src/Util/QueryHelper.ts
@@ -1,0 +1,7 @@
+import { OrderDirection } from "./types";
+
+export interface GetQueryOptions<T> {
+  orderBy?: Record<keyof T, OrderDirection>;
+  limit?: number;
+  offset?: number;
+}